requirements.txt
twilio~=7.3.0
python-dotenv~=0.19.2
fastapi~=0.70.0
uvicorn~=0.15.0
pymongo~=3.12.1